Rename factset.video.gui -> factset.gui.video (so that GUI could potentially be in its own jar someday)
