+++ factset-tooling/src/main/java/org/mytake/factset/gui/video/MismatchCtl.java
@@ -26,7 +26,7 @@
  *
  * You can contact us at team@mytake.org
  */
-package org.mytake.factset.video.gui;
+package org.mytake.factset.gui.video;
 
 
 import com.diffplug.common.base.Either;
@@ -46,6 +46,7 @@ import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Label;
 import org.eclipse.swt.widgets.Text;
+import org.mytake.factset.gui.Labels;
 import org.mytake.factset.video.TranscriptMatch;
 import org.mytake.factset.video.Word;
