+++ factset-tooling/src/main/java/org/mytake/factset/gui/video/MultiVideoSync.java
@@ -26,7 +26,7 @@
  *
  * You can contact us at team@mytake.org
  */
-package org.mytake.factset.video.gui;
+package org.mytake.factset.gui.video;
 
 
 import com.diffplug.common.swt.Layouts;
@@ -36,6 +36,7 @@ import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Text;
+import org.mytake.factset.gui.Labels;
 
 /** Tool to play back videos with an offset. */
 public class MultiVideoSync {
