--- factset-tooling/src/main/java/org/mytake/factset/video/gui/SaidVtt.java
@@ -1,47 +1,0 @@
-/*
- * MyTake.org website and tooling.
- * Copyright (C) 2018-2020 MyTake.org, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * Additional permission under GNU GPL version 3 section 7
- *
- * If you modify this Program, or any covered work, by linking or combining it
- * with Eclipse SWT (or a modified version of that library), containing parts
- * covered by the terms of the Eclipse Public License, the licensors of this Program
- * grant you additional permission to convey the resulting work.
- * {Corresponding Source for a non-source form of such a combination shall include the
- * source code for the parts of Eclipse SWT used as well as that of the covered work.}
- *
- * You can contact us at team@mytake.org
- */
-package org.mytake.factset.video.gui;
-
-
-import com.diffplug.common.base.Unhandled;
-
-public enum SaidVtt {
-	SAID, VTT;
-
-	public <T> T saidVtt(T said, T vtt) {
-		switch (this) {
-		case SAID:
-			return said;
-		case VTT:
-			return vtt;
-		default:
-			throw Unhandled.enumException(this);
-		}
-	}
-}
